requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
weaviate-client==4.6.6
python-dotenv==1.0.1
numpy==1.26.4
requests==2.32.3
pdfplumber==0.9.0
python-docx==0.8.11